Agregando conexión a bd y API

+++ Publicaciones/Publicaciones.Domain/Entities/Titles.cs
@@ -12,8 +12,8 @@ namespace Publicaciones.Domain.Entities
         public string Title { get; set; }
         public string Type { get; set; }
         public string? PubID { get; set; }
-        public double? Price { get; set; }
-        public double? Advance { get; set; }
+        public decimal? Price { get; set; }
+        public decimal? Advance { get; set; }
         public int? Royalty { get; set; }
         public int? Ytd_Sales { get; set; }
         public string? Notes { get; set; }
